# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Qlxion API
  description: |
    Qlxion API Documentation - API Gateway dan Auth Service.
    
    Semua endpoint yang dilindungi (Bearer Auth) memerlukan header:
    `Authorization: Bearer <access_token>`
    
    Gateway akan meneruskan claims JWT ke backend services melalui header:
    - `X-User-Id`: UUID user
    - `X-Tenant-Id`: UUID tenant
    - `X-User-Email`: Email user
    - `X-User-Roles`: Comma-separated roles
  version: 1.0.0
  contact:
    name: Qlxion Team
    email: dev@qlxion.com
servers:
  - url: http://localhost:8000
    description: Local development (API Gateway)
  - url: http://localhost:8001
    description: Local development (Auth Service direct)
security:
  - BearerAuth: []
paths:
...
  /roles:
    get:
      tags:
        - Roles
      summary: List roles
      description: Get paginated list of roles
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoleResponse'

    post:
      tags:
        - Roles
      summary: Create role
      description: Create a new role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created
  /roles/{id}:
    get:
      tags:
        - Roles
      summary: Get role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Role found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'

    put:
      tags:
        - Roles
      summary: Update role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated

    delete:
      tags:
        - Roles
      summary: Delete role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role deleted
  /roles/{id}/permissions:
    get:
      tags:
        - Roles
      summary: Get role permissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Role permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'

    post:
      tags:
        - Roles
      summary: Assign permission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permission_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Permission assigned

  # ============================================
  # Permissions
  # ============================================
  /permissions:
    get:
      tags:
        - Permissions
      summary: List permissions
      responses:
        '200':
          description: List of permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'

    post:
      tags:
        - Permissions
      summary: Create permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePermissionRequest'
      responses:
        '201':
          description: Permission created

  # ============================================
  # Tenants
  # ============================================
...
  /sessions:
    get:
      tags:
        - Sessions
      summary: List sessions
      parameters:
        - name: device_type
          in: query
          schema:
            type: string
            enum: [mobile, desktop, web]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, revoked]
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSessionResponse'
  /sessions/{id}/revoke:
    post:
      tags:
        - Sessions
      summary: Revoke session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session revoked

  # ============================================
  # Audit Logs
  # ============================================
  /audit-logs:
    get:
      tags:
        - Audit Logs
      summary: List audit logs
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: resource
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogResponse'
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ============================================
  # Schemas
  # ============================================
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
          example: ROLE_ADMIN
        name:
          type: string
          example: Administrator
        description:
          type: string
          nullable: true
        is_system_defined:
          type: boolean
        status:
          type: integer
          enum: [0, 1]
          description: 0=INACTIVE, 1=ACTIVE
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        created_at:
          type: string
          format: date-time
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        resource:
          type: string
          example: user
        action:
          type: string
          example: create
        description:
          type: string
          nullable: true
        is_system_defined:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        device_name:
          type: string
          nullable: true
        device_type:
          type: string
          enum: [mobile, desktop, web]
          nullable: true
        ip_address:
          type: string
          nullable: true
        user_agent:
          type: string
        expires_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          nullable: true
        last_activity_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
          nullable: true
        user_id:
          type: string
          format: uuid
          nullable: true
        action:
          type: string
          example: LOGIN
        resource:
          type: string
          nullable: true
        resource_id:
          type: string
          format: uuid
          nullable: true
        details:
          type: object
        ip_address:
          type: string
          nullable: true
        user_agent:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    # Request schemas
...
    CreateRoleRequest:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          example: ROLE_MANAGER
        name:
          type: string
        description:
          type: string
          nullable: true
        tenant_id:
          type: string
          format: uuid
          nullable: true
        permission_ids:
          type: array
          items:
            type: string
            format: uuid
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: integer
    CreatePermissionRequest:
      type: object
      required:
        - resource
        - action
      properties:
        resource:
          type: string
        action:
          type: string
        description:
          type: string
          nullable: true
...
    PaginatedRoleResponse:
      allOf:
        - $ref: '#/components/schemas/JSendSuccess'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Role'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
...
    PaginatedSessionResponse:
      allOf:
        - $ref: '#/components/schemas/JSendSuccess'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Session'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
    PaginatedAuditLogResponse:
      allOf:
        - $ref: '#/components/schemas/JSendSuccess'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AuditLog'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
    JSendSuccess:
      type: object
      properties:
        status:
          type: string
          example: success
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

  # ============================================
  # Responses
  # ============================================
...